feat: hide paid problem solution
